Add error boundary to tile components, propagate more errors
